@@ -1,11 +1,13 @@
 <Project Sdk="Microsoft.NET.Sdk">
 
     <PropertyGroup>
-        <TargetFramework>net5.0</TargetFramework>
+        <TargetFramework>net6.0</TargetFramework>
+        <LangVersion>10</LangVersion>
         <RootNamespace>Png</RootNamespace>
     </PropertyGroup>
 
     <ItemGroup>
-      <ProjectReference Include="..\CowLibrary\CowLibrary.csproj" />
+      <ProjectReference Include="..\Cowject\Cowject.csproj" />
+      <ProjectReference Include="..\ImageWorker\ImageWorker.csproj" />
     </ItemGroup>
 </Project>
