@@ -1,16 +1,18 @@
 <Project Sdk="Microsoft.NET.Sdk">
 
     <PropertyGroup>
-        <TargetFramework>net5.0</TargetFramework>
+        <TargetFramework>net6.0</TargetFramework>
+        <LangVersion>10</LangVersion>
     </PropertyGroup>
 
     <ItemGroup>
-      <PackageReference Include="SceneFormat" Version="0.1.181" />
+      <PackageReference Include="SceneFormat" Version="0.1.216" />
     </ItemGroup>
 
     <ItemGroup>
       <ProjectReference Include="..\Cowject\Cowject.csproj" />
       <ProjectReference Include="..\CowRenderer\CowRenderer.csproj" />
+      <ProjectReference Include="..\ObjWorker\ObjWorker.csproj" />
     </ItemGroup>
 
 </Project>
