@@ -10,24 +10,29 @@ namespace CowEngine
 
     public class ArgumentsParser : IArgumentsParser
     {
-        [Inject(Name = Options.Model)]
-        public IFlow ModelFlow { get; set; }
-
-        [Inject(Name = Options.Compiled)]
-        public IFlow CompiledFlow { get; set; }
-
-        [Inject(Name = Options.Scene)]
-        public IFlow SceneFlow { get; set; }
+        [Inject]
+        public IFlow[] Flows { get; set; }
 
         public void Parse(string[] args)
         {
-            Parser.Default.ParseArguments<ModelOptions, CompiledOptions, SceneOptions>(args)
-                .MapResult<ModelOptions, CompiledOptions, SceneOptions, int>(
-                    opts => ModelFlow.Process(opts.Source, opts.Output),
-                    opts => CompiledFlow.Process("", opts.Output),
-                    opts => SceneFlow.Process(opts.Source, opts.Output),
+            Parser.Default
+                .ParseArguments<Option>(args)
+                .MapResult(
+                    Process,
                     _ => 1
                 );
+        }
+
+        private int Process(Option option)
+        {
+            foreach (var flow in Flows)
+            {
+                if (flow.CanWorkWithProcess(option))
+                {
+                    return flow.Process(option);
+                }
+            }
+            return 1;
         }
     }
 }
