@@ -5,14 +5,19 @@ namespace CowEngine
 
     public class CoreModule : IModule
     {
+        public Priority Priority => Priority.Highest;
+
         public void Prepare(DiContainer container)
         {
-            container.Bind<IArgumentsParser>().To<ArgumentsParser>().ToSingleton();
+            container.Bind<IBytesReader>().To<SimpleBytesReader>().ToSingleton();
+            container.Bind<IBytesWriter>().ToInstance(new FilesBypassingWriter("_"));
             container.Bind<IIoWorker>().To<IoWorker>().ToSingleton();
-            container.Bind<IWatch>().To<Watch>().ToSingleton();
-            container.Bind<IFlow>().WithName(Options.Model).To<ModelFlow>().ToSingleton();
-            container.Bind<IFlow>().WithName(Options.Compiled).To<CompiledFlow>().ToSingleton();
-            container.Bind<IFlow>().WithName(Options.Scene).To<SceneFlow>().ToSingleton();
+            
+            container.BindInterfacesTo<ArgumentsParser>().ToSingleton();
+            container.BindInterfacesTo<Watch>().ToSingleton();
+            container.BindInterfacesTo<SceneLoader>().ToSingleton();
+            container.BindInterfacesTo<BasicFlow>().ToSingleton();
+            container.BindInterfacesTo<RendererProvider>().ToSingleton();
         }
     }
 }
