@@ -2,6 +2,8 @@ namespace CowEngine
 {
     public interface IFlow
     {
-        public int Process(string source, string output);
+        public bool CanWorkWithProcess(Option option);
+
+        public int Process(Option option);
     }
 }
