@@ -37,8 +37,7 @@ namespace CowLibrary
 
         public static Vector3 Reflect(this Vector3 v, Vector3 n)
         {
-            var num = -2f * Vector3.Dot(n, v);
-            return new Vector3(num * n.X + v.X, num * n.Y + v.Y, num * n.Z + v.Z);
+            return -v + 2 * Vector3.Dot(v, n) * n;
         }
 
         public static Vector3 Faceforward(Vector3 n, Vector3 v)
@@ -50,7 +49,7 @@ namespace CowLibrary
         {
             var cosThetaI = Vector3.Dot(n, v);
             var sin2ThetaI = Math.Max(0, 1 - cosThetaI * cosThetaI);
-            var sin2ThetaT = eta * eta * sin2ThetaI;
+            var sin2ThetaT = eta * eta * sin2ThetaI; //todo check
             if (sin2ThetaT >= 1)
             {
                 w = Vector3.Zero;
