@@ -4,8 +4,10 @@ namespace CowLibrary
 
     public interface IBrdf
     {
-        public float Evaluate(Vector3 wo, Vector3 wi);
+        public float Evaluate(in Vector3 wo, in Vector3 wi);
 
-        public float Sample(Surfel surfel, out Vector3 wi, Vector2 sample, out float pdf);
+        public float Sample(in Vector3 wo, in Vector2 sample, out Vector3 wi, out float pdf);
+        
+        public float Pdf(in Vector3 wo, in Vector3 wi);
     }
 }
