@@ -3,29 +3,32 @@ namespace CowLibrary
     using System;
     using System.Numerics;
 
-    public class SpecularReflectionBrdf : IBrdf
+    public readonly struct SpecularReflectionBrdf : IBrdf
     {
         private readonly float r;
-        private readonly Fresnel fresnel;
+        private readonly DielectricFresnel fresnel;
 
-        public SpecularReflectionBrdf(float r, Fresnel fresnel)
+        public SpecularReflectionBrdf(float r, DielectricFresnel fresnel)
         {
             this.r = r;
             this.fresnel = fresnel;
         }
 
-        public float Evaluate(Vector3 wo, Vector3 wi)
+        public float Evaluate(in Vector3 wo, in Vector3 wi)
         {
             return 0;
         }
 
-        public float Sample(Surfel surfel, out Vector3 wi, Vector2 sample, out float pdf)
+        public float Sample(in Vector3 wo, in Vector2 sample, out Vector3 wi, out float pdf)
         {
             pdf = 1;
-            var wo = surfel.ray;
-            wi = wo.Reflect(surfel.normal);
-            var cos = Vector3.Dot(wi, surfel.normal);
-            return fresnel.Evaluate(cos) * r / Math.Abs(cos);
+            wi = new Vector3(-wo.X, wo.Y, -wo.Z);
+            return fresnel.Evaluate(Mathf.CosTheta(wi)) * r / Mathf.AbsCosTheta(wi);
+        }
+
+        public float Pdf(in Vector3 wo, in Vector3 wi)
+        {
+            return 0;
         }
     }
 }
