@@ -2,19 +2,24 @@ namespace CowLibrary
 {
     using System;
 
-    public class DielectricFresnel : Fresnel
+    public readonly struct DielectricFresnel : IFresnel
     {
-        public DielectricFresnel(float etaI, float etaT) : base(etaI, etaT)
+        private readonly float etaI;
+        private readonly float etaT;
+
+        public DielectricFresnel(float etaI, float etaT)
         {
+            this.etaI = etaI;
+            this.etaT = etaT;
         }
 
-        public override float Evaluate(float cosThetaI)
+        public float Evaluate(float cosThetaI)
         {
-            cosThetaI = Math.Clamp(cosThetaI, -1, 1);
+            cosThetaI = Mathf.Clamp(cosThetaI, -1, 1);
             var etaI = this.etaI;
             var etaT = this.etaT;
             var entering = cosThetaI > 0;
-            if (entering)
+            if (!entering)
             {
                 (etaI, etaT) = Mathf.Swap(etaI, etaT);
                 cosThetaI = Math.Abs(cosThetaI);
@@ -22,12 +27,16 @@ namespace CowLibrary
 
             var sinThetaI = Math.Sqrt(Math.Max(0, 1 - cosThetaI * cosThetaI));
             var sinThetaT = etaI / etaT * sinThetaI;
-            if (sinThetaT >= 0) return 1;
+            if (sinThetaT >= 1) return 1;
 
             var cosThetaT = Math.Sqrt(Math.Max(0, 1 - sinThetaT * sinThetaT));
 
-            var rParl = (etaT * cosThetaI - etaI * cosThetaT) / (etaT * cosThetaI + etaI * cosThetaT);
-            var rPerp = (etaI * cosThetaI - etaT * cosThetaT) / (etaI * cosThetaI + etaT * cosThetaT);
+            var rParl =
+                (etaI * cosThetaI - etaT * cosThetaT) /
+                (etaI * cosThetaI + etaT * cosThetaT);
+            var rPerp =
+                (etaI * cosThetaT - etaT * cosThetaI) /
+                (etaI * cosThetaT + etaT * cosThetaI);
 
             return (float)(rParl * rParl + rPerp * rPerp) * 0.5f;
         }
