@@ -1,16 +1,8 @@
 namespace CowLibrary
 {
-    public class FresnelNoOp : Fresnel
+    public readonly struct FresnelNoOp : IFresnel
     {
-        public FresnelNoOp() : base(1, 1)
-        {
-        }
-
-        public FresnelNoOp(float etaI, float etaT) : base(etaI, etaT)
-        {
-        }
-
-        public override float Evaluate(float cosThetaI)
+        public float Evaluate(float cosThetaI)
         {
             return 1;
         }
