@@ -1,31 +1,15 @@
 namespace CowLibrary.Lights
 {
-    using System.Numerics;
+    using Mathematics.Sampler;
+    using Models;
 
     public class DirectionalLight : Light
     {
-        private readonly Color color;
-        private readonly float intensity;
+        public override ILightModel Model { get; }
 
-        public DirectionalLight(Color color, float intensity)
+        public DirectionalLight(Color color, float intensity, int id, ISampler sampler) : base(sampler)
         {
-            this.color = color;
-            this.intensity = intensity;
-        }
-
-        public override ShadingInfo GetShadingInfo(Surfel surfel)
-        {
-            return new ShadingInfo()
-            {
-                direction = Transform.Forward,
-                distance = float.PositiveInfinity,
-                color = color * intensity
-            };
-        }
-
-        public override Color Sample(Vector3 wi)
-        {
-            return Color.Black;
+            Model = new DirectionalLightModel(color * intensity, id);
         }
     }
 }
