@@ -1,32 +1,15 @@
 namespace CowLibrary.Lights
 {
-    using System;
-    using System.Numerics;
+    using Mathematics.Sampler;
+    using Models;
 
     public class EnvironmentLight : Light
     {
-        private readonly Color color;
-        private readonly float intensity;
+        public override ILightModel Model { get; }
 
-        public EnvironmentLight(Color color, float intensity)
+        public EnvironmentLight(Color color, float intensity, int id, ISampler sampler) : base(sampler)
         {
-            this.color = color;
-            this.intensity = intensity;
-        }
-
-        public override ShadingInfo GetShadingInfo(Surfel surfel)
-        {
-            return new ShadingInfo()
-            {
-                direction = Mathf.CosineSampleHemisphere(surfel.normal, Mathf.CreateSample()),
-                distance = float.PositiveInfinity,
-                color = color * intensity
-            };
-        }
-
-        public override Color Sample(Vector3 wi)
-        {
-            return color * intensity;
+            Model = new EnvironmentLightModel(color * intensity, id);
         }
     }
 }
