@@ -1,35 +1,15 @@
 namespace CowLibrary.Lights
 {
-    using System;
-    using System.Numerics;
+    using Mathematics.Sampler;
+    using Models;
 
     public class PointLight : Light
     {
-        private readonly Color color;
-        private readonly float intensity;
+        public override ILightModel Model { get; }
 
-        public PointLight(Color color, float intensity)
+        public PointLight(Color color, float intensity, int id, ISampler sampler) : base(sampler)
         {
-            this.color = color;
-            this.intensity = intensity;
-        }
-
-        public override ShadingInfo GetShadingInfo(Surfel surfel)
-        {
-            var direction = Transform.Position - surfel.point;
-            var sqrtDistance = direction.LengthSquared();
-            var distance = (float)Math.Sqrt(sqrtDistance);
-            return new ShadingInfo
-            {
-                direction = direction / distance,
-                distance = distance,
-                color = color * (intensity / (4 * Math.PI * sqrtDistance))
-            };
-        }
-
-        public override Color Sample(Vector3 wi)
-        {
-            return Color.Black;
+            Model = new PointLightModel(color * intensity, id);
         }
     }
 }
