@@ -2,23 +2,38 @@ namespace CowLibrary
 {
     using System.Numerics;
 
-    public class DiffuseMaterial : Material
+    public readonly struct DiffuseMaterial : IMaterial
     {
-        private readonly IBrdf brdf;
+        public Color Color { get; }
 
-        public DiffuseMaterial(Color color, float r) : base(color)
+        public int Id { get; }
+
+        private readonly LambertianBrdf brdf;
+
+        public DiffuseMaterial(Color color, float r, int id) : this(color, new LambertianBrdf(r), id)
         {
-            brdf = new LambertianBrdf(r);
         }
 
-        public override Color GetColor(Vector3 wo, Vector3 wi)
+        private DiffuseMaterial(Color color, LambertianBrdf brdf, int id)
         {
-            return brdf.Evaluate(wo, wi) * Color;
+            Color = color;
+            this.brdf = brdf;
+            Id = id;
         }
 
-        public override float Sample(Surfel surfel, out Vector3 wi, out float pdf)
+        public Color GetColor(in Vector3 wo, in Vector3 wi)
         {
-            return brdf.Sample(surfel, out wi, Mathf.CreateSample(), out pdf);
+            return brdf.Evaluate(in wo, in wi) * Color;
+        }
+
+        public Color Sample(in Vector3 wo, in Vector2 sample, out Vector3 wi, out float pdf)
+        {
+            return brdf.Sample(in wo, in sample, out wi, out pdf) * Color;
+        }
+
+        public IMaterial Copy(int id)
+        {
+            return new DiffuseMaterial(Color, brdf, id);
         }
     }
 }
