@@ -2,24 +2,39 @@ namespace CowLibrary
 {
     using System.Numerics;
 
-    public class ReflectionMaterial : Material
+    public readonly struct ReflectionMaterial : IMaterial
     {
-        private IBrdf brdf;
+        public Color Color { get; }
 
-        public ReflectionMaterial(float r, float eta) : base(Color.White)
+        public int Id { get; }
+
+        private readonly SpecularReflectionBrdf brdf;
+
+        public ReflectionMaterial(Color color, float r, float eta, int id) :
+            this(color, new SpecularReflectionBrdf(r, new DielectricFresnel(1, eta)), id)
         {
-            var fresnel = new DielectricFresnel(1, eta);
-            brdf = new SpecularReflectionBrdf(r, fresnel);
         }
 
-        public override Color GetColor(Vector3 wo, Vector3 wi)
+        private ReflectionMaterial(Color color, SpecularReflectionBrdf brdf, int id)
         {
-            return brdf.Evaluate(wo, wi) * Color;
+            Color = color;
+            this.brdf = brdf;
+            Id = id;
         }
 
-        public override float Sample(Surfel surfel, out Vector3 wi, out float pdf)
+        public Color GetColor(in Vector3 wo, in Vector3 wi)
         {
-            return brdf.Sample(surfel, out wi, Mathf.CreateSample(), out pdf);
+            return brdf.Evaluate(in wo, in wi) * Color;
+        }
+
+        public Color Sample(in Vector3 wo, in Vector2 sample, out Vector3 wi, out float pdf)
+        {
+            return brdf.Sample(in wo, in sample, out wi, out pdf) * Color;
+        }
+
+        public IMaterial Copy(int id)
+        {
+            return new ReflectionMaterial(Color, brdf, id);
         }
     }
 }
