@@ -1,55 +1,41 @@
 namespace CowLibrary
 {
     using System.Numerics;
+    using Object.Mesh.Views;
 
-    public class Plane : Mesh
+    public struct Plane : IMesh<PlaneView>
     {
-        private const float e = 1e-10f;
+        public readonly int Id => view.Id;
 
-        public override Box BoundingBox => box;
+        public readonly PlaneView View => view;
 
-        private Box box;
+        public readonly Bound BoundingBox => bound;
 
-        private Vector3 normal = -Vector3.UnitY;
-        private Vector3 point = Vector3.Zero;
+        private Bound bound;
+        private PlaneView view;
 
-        public Plane()
+        public Plane(int id) : this()
         {
-            box = CreateBox();
+            view = new PlaneView(Vector3.Zero, Vector3.UnitY, id);
+            bound = CreateBound();
         }
 
-        private Box CreateBox()
+        private Bound CreateBound()
         {
-            return new Box(point, 1000);
+            return new Bound(view.point, 1000, Id);
         }
 
-        public override bool Intersect(Ray ray, out Surfel surfel)
+        public readonly void Intersect(in Ray ray, ref RayHit best)
         {
-            var dot = Vector3.Dot(normal, ray.direction);
-            if (dot > e)
-            {
-                var dir = point - ray.origin;
-                var t = Vector3.Dot(dir, normal) / dot;
-                if (t > 0)
-                {
-                    surfel = new Surfel()
-                    {
-                        t = t,
-                        point = ray.GetPoint(t),
-                        normal = -normal
-                    };
-                    return true;
-                }
-            }
-            surfel = null;
-            return false;
+            view.Intersect(in ray, ref best);
         }
 
-        public override void Apply(Matrix4x4 matrix)
+        public void Apply(in Matrix4x4 matrix)
         {
-            normal = matrix.MultiplyVector(normal).Normalize();
-            point = matrix.MultiplyPoint(point);
-            box = CreateBox();
+            var point = matrix.MultiplyPoint(view.point);
+            var normal = matrix.MultiplyVector(view.normal).Normalize();
+            view = new PlaneView(point, normal, Id);
+            bound = CreateBound();
         }
     }
 }
