@@ -2,113 +2,52 @@ namespace CowLibrary
 {
     using System;
     using System.Numerics;
+    using Views;
 
-    public class Triangle : Mesh
+    public struct Triangle : IMesh<TriangleView>
     {
-        private const double e = 1e-10f;
+        public readonly int Id => view.Id;
 
-        public override Box BoundingBox => box;
+        public readonly TriangleView View => view;
 
-        public Vector3 v0;
-        public Vector3 v1;
-        public Vector3 v2;
+        public readonly Bound BoundingBox => bound;
+        
+        private TriangleView view;
+        private Bound bound;
 
-        public Vector3 n0;
-        public Vector3 n1;
-        public Vector3 n2;
-
-        private Box box;
-
-        public Triangle(Vector3 v0, Vector3 v1, Vector3 v2)
+        public Triangle(Vector3 v0, Vector3 v1, Vector3 v2, Vector3 n0, Vector3 n1, Vector3 n2, int id) : this()
         {
-            this.v0 = v0;
-            this.v1 = v1;
-            this.v2 = v2;
-            box = CreateBox();
+            view = new TriangleView(v0, v1, v2, n0, n1, n2, id);
+            bound = CreateBound();
         }
 
-        private Box CreateBox()
+        private Bound CreateBound()
         {
             Vector3 min, max;
-            min.X = Math.Min(v0.X, Math.Min(v1.X, v2.X));
-            min.Y = Math.Min(v0.Y, Math.Min(v1.Y, v2.Y));
-            min.Z = Math.Min(v0.Z, Math.Min(v1.Z, v2.Z));
-            max.X = Math.Max(v0.X, Math.Max(v1.X, v2.X));
-            max.Y = Math.Max(v0.Y, Math.Max(v1.Y, v2.Y));
-            max.Z = Math.Max(v0.Z, Math.Max(v1.Z, v2.Z));
-            return new Box(min, max);
+            min.X = Math.Min(view.v0.X, Math.Min(view.v1.X, view.v2.X));
+            min.Y = Math.Min(view.v0.Y, Math.Min(view.v1.Y, view.v2.Y));
+            min.Z = Math.Min(view.v0.Z, Math.Min(view.v1.Z, view.v2.Z));
+            max.X = Math.Max(view.v0.X, Math.Max(view.v1.X, view.v2.X));
+            max.Y = Math.Max(view.v0.Y, Math.Max(view.v1.Y, view.v2.Y));
+            max.Z = Math.Max(view.v0.Z, Math.Max(view.v1.Z, view.v2.Z));
+            return new Bound(min, max, Id);
         }
 
-        public void SetNormal(Vector3 n0, Vector3 n1, Vector3 n2)
+        public readonly void Intersect(in Ray ray, ref RayHit best)
         {
-            this.n0 = n0.Normalize();
-            this.n1 = n1.Normalize();
-            this.n2 = n2.Normalize();
+            view.Intersect(in ray, ref best);
         }
 
-        public void CalculateNormal()
+        public void Apply(in Matrix4x4 matrix)
         {
-            var v0v1 = v1 - v0;
-            var v0v2 = v2 - v0;
-            var n = Vector3.Cross(v0v2, v0v1);
-            n0 = n1 = n2 = n;
-        }
-
-        public override bool Intersect(Ray ray, out Surfel surfel)
-        {
-            var edge1 = v1 - v0;
-            var edge2 = v2 - v0;
-
-            var h = Vector3.Cross(ray.direction, edge2);
-            var a = Vector3.Dot(edge1, h);
-            if (Math.Abs(a) < e)
-            {
-                surfel = null;
-                return false;
-            }
-
-            var f = 1f / a;
-            var s = ray.origin - v0;
-            var u = f * Vector3.Dot(s, h);
-            if (u < 0 || u > 1)
-            {
-                surfel = null;
-                return false;
-            }
-
-            var q = Vector3.Cross(s, edge1);
-            var v = f * Vector3.Dot(ray.direction, q);
-            if (v < 0 || u + v > 1)
-            {
-                surfel = null;
-                return false;
-            }
-
-            var t = f * Vector3.Dot(edge2, q);
-            if (t <= 0)
-            {
-                surfel = null;
-                return false;
-            }
-
-            surfel = new Surfel()
-            {
-                point = ray.GetPoint(t),
-                normal = n0 * (1 - u - v) + n1 * u + n2 * v,
-                t = t
-            };
-            return true;
-        }
-
-        public override void Apply(Matrix4x4 matrix)
-        {
-            v0 = matrix.MultiplyPoint(v0);
-            v1 = matrix.MultiplyPoint(v1);
-            v2 = matrix.MultiplyPoint(v2);
-            n0 = matrix.MultiplyVector(n0).Normalize();
-            n1 = matrix.MultiplyVector(n1).Normalize();
-            n2 = matrix.MultiplyVector(n2).Normalize();
-            box = CreateBox();
+            var v0 = matrix.MultiplyPoint(view.v0);
+            var v1 = matrix.MultiplyPoint(view.v1);
+            var v2 = matrix.MultiplyPoint(view.v2);
+            var n0 = matrix.MultiplyVector(view.n0).Normalize();
+            var n1 = matrix.MultiplyVector(view.n1).Normalize();
+            var n2 = matrix.MultiplyVector(view.n2).Normalize();
+            view = new TriangleView(v0, v1, v2, n0, n1, n2, Id);
+            bound = CreateBound();
         }
     }
 }
