@@ -2,15 +2,13 @@ namespace CowRenderer
 {
     public class RenderConfig
     {
-        public int width = 1920 / 3;
-        public int height = 1080 / 3;
+        public int width = 1920;
+        public int height = 1080;
         public int fov = 60;
 
-        public float bias = 0.00001f;
-
         public int numberOfThreadPerDimension = 8;
-        public int numberOfRayPerPixel = 16;
-        public int rayDepth = 2;
-        public int numberOfRayPerLight = 2;
+        public int numberOfRayPerPixelDimension = 8;
+        public int rayDepth = 3;
+        public int numberOfRayPerMaterial = 1;
     }
 }
