@@ -6,6 +6,7 @@ namespace CowRenderer
     using System.Numerics;
     using Cowject;
     using CowLibrary;
+    using CowLibrary.Mathematics.Sampler;
 
     public class AutoAdjustScene : Scene
     {
@@ -14,25 +15,32 @@ namespace CowRenderer
 
         public override Camera MainCamera => camera;
 
+        private readonly ISampler sampler;
         private RealisticCamera camera;
+
+        public AutoAdjustScene(RenderableObject obj, ISampler sampler)
+        {
+            objects.Add(obj);
+            this.sampler = sampler;
+        }
 
         public override void PrepareScene()
         {
             camera = CreateCamera();
             cameras.Add(camera);
-            var box = GetBoundingBoxFor(objects);
-            PlaceCamera(box);
-            PlacePlane(box);
+            var bound = GetBoundingBoxFor(objects);
+            PlaceCamera(bound);
+            PlacePlane(bound);
             base.PrepareScene();
         }
 
         private RealisticCamera CreateCamera()
         {
-            return new RealisticCamera(RenderConfig.width, RenderConfig.height, RenderConfig.fov,
+            return new RealisticCamera(RenderConfig.width, RenderConfig.height, sampler, RenderConfig.fov,
                 new Lens(1f, 0.01f, 1f));
         }
 
-        private Box GetBoundingBoxFor(List<RenderableObject> renderableObjects)
+        private Bound GetBoundingBoxFor(List<RenderableObject> renderableObjects)
         {
             var min = renderableObjects.First().Mesh.BoundingBox.min;
             var max = renderableObjects.First().Mesh.BoundingBox.max;
@@ -47,21 +55,21 @@ namespace CowRenderer
                 max.Z = Math.Max(max.Z, objectBoundingBox.max.Z);
             }
 
-            return new Box(min, max);
+            return new Bound(min, max, -1);
         }
 
-        private void PlaceCamera(Box box)
+        private void PlaceCamera(Bound box)
         {
-            var max = Math.Max(box.size.X * 1.3f, box.size.Y * 1.3f * camera.AspectRatio);
+            var max = Math.Max(box.Size.X * 1.3f, box.Size.Y * 1.3f * camera.AspectRatio);
             var tan = (float)Math.Tan(Const.Deg2Rad * camera.Fov / 2);
             var dist = max / tan;
 
-            camera.Transform.LocalToWorldMatrix = Matrix4x4Extensions.LookAt(new Vector3(0, 0.5f, dist), box.center);
+            camera.Transform.LocalToWorldMatrix = Matrix4x4Extensions.LookAt(new Vector3(0, 0.5f, dist), box.Center);
         }
 
-        private void PlacePlane(Box box)
+        private void PlacePlane(Bound box)
         {
-            var plane = new RenderableObject(new Disk(100), new DiffuseMaterial(Color.Red, 1));
+            var plane = new RenderableObject(new Disk(100, 1), new DiffuseMaterial(Color.Red, 1, 1));
             plane.Transform.Position = box.min.Y * Vector3.UnitY;
             objects.Add(plane);
         }
