@@ -1,10 +1,8 @@
 namespace CowEngine.ImageWorker
 {
     using System;
-    using System.Collections.Generic;
     using System.Linq;
     using Cowject;
-    using Png;
 
     internal interface IImageEncoderProvider
     {
@@ -13,14 +11,12 @@ namespace CowEngine.ImageWorker
 
     public class ImageEncoderProvider : IImageEncoderProvider
     {
-        private readonly List<IImageEncoder> encoders = new()
-        {
-            new PngWorker()
-        };
+        [Inject]
+        public IImageEncoder[] Encoders { get; set; }
 
         public IImageEncoder FindEncoder(string extension)
         {
-            var encoder = encoders.FirstOrDefault(en => en.CanWorkWith(extension));
+            var encoder = Encoders.FirstOrDefault(en => en.CanWorkWith(extension));
             if (encoder == null)
             {
                 throw new Exception("Cannot encode image. Format is not supported");
