@@ -1,6 +1,5 @@
 namespace CowEngine
 {
-    using System.Collections.Generic;
     using System.Numerics;
     using CowLibrary;
     using ObjLoader.Loader.Data.Elements;
@@ -9,35 +8,46 @@ namespace CowEngine
 
     public interface IModelToObjectConverter
     {
-        Mesh Convert(LoadResult result);
+        IMesh Convert(LoadResult result, int id);
     }
 
     public class ModelToMeshConverter : IModelToObjectConverter
     {
-        public Mesh Convert(LoadResult result)
+        public IMesh Convert(LoadResult result, int id)
         {
-            return ExtractMesh(result, result.Groups[0]);
+            return ExtractMesh(result, result.Groups[0], id);
         }
 
-        private Mesh ExtractMesh(LoadResult result, Group group)
+        private IMesh ExtractMesh(LoadResult result, Group group, int id)
         {
-            var triangles = new List<Triangle>();
-            foreach (var face in group.Faces)
+            var triangles = new Triangle[group.Faces.Count];
+            for (var i = 0; i < group.Faces.Count; i++)
             {
+                var face = group.Faces[i];
                 var (v0, v1, v2) = GetVertices(result, face);
-                var t = new Triangle(v0, v1, v2);
+                Vector3 n0;
+                Vector3 n1;
+                Vector3 n2;
                 if (result.Normals.Count == 0)
                 {
-                    t.CalculateNormal();
+                    (n0, n1, n2) = CalculateNormal(v0, v1, v2);
                 }
                 else
                 {
-                    var (n0, n1, n2) = GetNormals(result, face);
-                    t.SetNormal(n0, n1, n2);
+                    (n0, n1, n2) = GetNormals(result, face);
                 }
-                triangles.Add(t);
+                var t = new Triangle(v0, v1, v2, n0, n1, n2, id);
+                triangles[i] = t;
             }
-            return new OptimizedMesh(triangles);
+            return new OptimizedMesh(triangles, id);
+        }
+
+        public (Vector3 n0, Vector3 n1, Vector3 n2) CalculateNormal(Vector3 v0, Vector3 v1, Vector3 v2)
+        {
+            var v0v1 = v1 - v0;
+            var v0v2 = v2 - v0;
+            var n = Vector3.Cross(v0v2, v0v1);
+            return (n, n, n);
         }
 
         private (Vector3 v0, Vector3 v1, Vector3 v2) GetVertices(LoadResult result, Face face)
