@@ -5,7 +5,7 @@ namespace CowEngine
 
     public interface IRenderableObjectWorker
     {
-        RenderableObject Parse(string source);
+        RenderableObject Parse(string source, int id);
     }
 
     public class RenderableObjectWorker : IRenderableObjectWorker
@@ -13,10 +13,10 @@ namespace CowEngine
         [Inject]
         public IObjWorker ObjWorker { get; set; }
 
-        public RenderableObject Parse(string source)
+        public RenderableObject Parse(string source, int id)
         {
-            var mesh = ObjWorker.Parse(source);
-            var material = new DiffuseMaterial(new Color(127, 0, 255), 1f);
+            var mesh = ObjWorker.Parse(source, id);
+            var material = new DiffuseMaterial(new Color(127, 0, 255), 1f, id);
             return new RenderableObject(mesh, material);
         }
     }
