@@ -8,22 +8,10 @@
     using Cowject;
     using CowLibrary;
     using CowLibrary.Lights;
+    using CowLibrary.Mathematics.Sampler;
     using CowRenderer;
     using Google.Protobuf.Collections;
-    using SceneFormat;
-    using Camera = CowLibrary.Camera;
-    using OrthographicCamera = CowLibrary.OrthographicCamera;
-    using PerspectiveCamera = CowLibrary.PerspectiveCamera;
-    using Scene = CowRenderer.Scene;
-    using Sphere = CowLibrary.Sphere;
     using Plane = CowLibrary.Plane;
-    using Disk = CowLibrary.Disk;
-    using Light = CowLibrary.Lights.Light;
-    using Color = CowLibrary.Color;
-    using Material = CowLibrary.Material;
-    using SceneObject = SceneFormat.SceneObject;
-    using Transform = CowLibrary.Transform;
-    using Vector3 = System.Numerics.Vector3;
 
     public interface ISceneWorker
     {
@@ -33,10 +21,13 @@
     public class SceneWorker : ISceneWorker
     {
         [Inject]
-        public ISceneIO SceneIo { get; set; }
+        public SceneFormat.ISceneIO SceneIo { get; set; }
 
         [Inject]
         public IObjWorker ObjWorker { get; set; }
+
+        [Inject]
+        public ISamplerProvider SamplerProvider { get; set; }
 
         [Inject]
         public RenderConfig RenderConfig { get; set; }
@@ -50,14 +41,22 @@
         private Scene ConvertToScene(SceneFormat.Scene parsedScene)
         {
             var scene = new ComplexScene();
-            RenderConfig.width = parsedScene.RenderOptions.Width;
-            RenderConfig.height = parsedScene.RenderOptions.Height;
+            ParseRenderOptions(parsedScene.RenderOptions);
             scene.cameras.AddRange(parsedScene.Cameras.Select(ConvertToCamera));
             scene.SetMainCamera(parsedScene.RenderOptions.CameraId);
             var materials = ParseMaterials(parsedScene.Materials);
-            scene.objects.AddRange(parsedScene.SceneObjects.Select(obj => ConvertToObject(obj, materials)));
+            scene.objects.AddRange(
+                parsedScene.SceneObjects.Select((obj, index) => ConvertToObject(obj, materials, index)));
             scene.lights.AddRange(parsedScene.Lights.Select(ConvertToLight));
             return scene;
+        }
+
+        private void ParseRenderOptions(SceneFormat.RenderOptions renderOptions)
+        {
+            RenderConfig.width = renderOptions.Width;
+            RenderConfig.height = renderOptions.Height;
+            RenderConfig.rayDepth = renderOptions.RayDepth;
+            RenderConfig.numberOfRayPerPixelDimension = renderOptions.RaysPerPixelDimension;
         }
 
         private Camera ConvertToCamera(SceneFormat.Camera parsedCamera)
@@ -66,21 +65,25 @@
             switch (parsedCamera.CameraCase)
             {
                 case SceneFormat.Camera.CameraOneofCase.Perspective:
-                    return new PerspectiveCamera()
+                    return new PerspectiveCamera(RenderConfig.width, RenderConfig.height, SamplerProvider.Sampler,
+                        (float)parsedCamera.Perspective.Fov)
                     {
                         Id = parsedCamera.Id,
                         Transform = tran,
-                        width = RenderConfig.width,
-                        height = RenderConfig.height,
-                        Fov = (float)parsedCamera.Perspective.Fov,
                     };
                 case SceneFormat.Camera.CameraOneofCase.Orthographic:
-                    return new OrthographicCamera()
+                    return new OrthographicCamera(RenderConfig.width, RenderConfig.height, SamplerProvider.Sampler)
                     {
                         Id = parsedCamera.Id,
                         Transform = tran,
-                        width = RenderConfig.width,
-                        height = RenderConfig.height,
+                    };
+                case SceneFormat.Camera.CameraOneofCase.Realistic:
+                    var lens = new Lens(1, (float)parsedCamera.Realistic.Radius, (float)parsedCamera.Realistic.Focus);
+                    return new RealisticCamera(RenderConfig.width, RenderConfig.height, SamplerProvider.Sampler,
+                        (float)parsedCamera.Realistic.Fov, lens)
+                    {
+                        Id = parsedCamera.Id,
+                        Transform = tran,
                     };
                 default:
                     throw new Exception("Unsupported camera");
@@ -88,11 +91,11 @@
         }
 
         private RenderableObject ConvertToObject(SceneFormat.SceneObject parsedObject,
-            Dictionary<string, Material> materials)
+            Dictionary<string, IMaterial> materials, int id)
         {
             var tran = ConvertTransform(parsedObject.Transform);
-            var mesh = GetMesh(parsedObject);
-            var material = GetMaterial(parsedObject, materials);
+            var mesh = GetMesh(parsedObject, id);
+            var material = GetMaterial(parsedObject, materials, id);
             return new RenderableObject(mesh, material)
             {
                 Id = parsedObject.Id,
@@ -100,62 +103,63 @@
             };
         }
 
-        private Mesh GetMesh(SceneFormat.SceneObject parsedObject)
+        private IMesh GetMesh(SceneFormat.SceneObject parsedObject, int id)
         {
             switch (parsedObject.MeshCase)
             {
                 case SceneFormat.SceneObject.MeshOneofCase.Sphere:
-                    return new Sphere((float)parsedObject.Sphere.Radius);
+                    return new Sphere((float)parsedObject.Sphere.Radius, id);
                 case SceneFormat.SceneObject.MeshOneofCase.Cube:
-                    return new Box(ConvertVector(parsedObject.Cube.Size) * 0.5f);
+                    return new Box(ConvertVector(parsedObject.Cube.Size) * 0.5f, id);
                 case SceneFormat.SceneObject.MeshOneofCase.Plane:
-                    return new Plane();
+                    return new Plane(id);
                 case SceneFormat.SceneObject.MeshOneofCase.Disk:
-                    return new Disk((float)parsedObject.Disk.Radius);
+                    return new Disk((float)parsedObject.Disk.Radius, id);
                 case SceneFormat.SceneObject.MeshOneofCase.MeshedObject:
-                    return ObjWorker.Parse(parsedObject.MeshedObject.Reference);
+                    return ObjWorker.Parse(parsedObject.MeshedObject.Reference, id);
                 default:
                     throw new Exception("Unsupported mesh");
             }
         }
 
-        private Material GetMaterial(SceneObject parsedObject, Dictionary<string, Material> materials)
+        private IMaterial GetMaterial(SceneFormat.SceneObject parsedObject, Dictionary<string, IMaterial> materials,
+            int id)
         {
             switch (parsedObject.ObjectMaterialCase)
             {
-                case SceneObject.ObjectMaterialOneofCase.MaterialId:
+                case SceneFormat.SceneObject.ObjectMaterialOneofCase.MaterialId:
                     if (materials.TryGetValue(parsedObject.MaterialId, out var material))
                     {
-                        return material;
+                        return material.Copy(id);
                     }
                     throw new Exception("Invalid material id");
-                case SceneObject.ObjectMaterialOneofCase.Material:
-                    return ConvertMaterial(parsedObject.Material);
+                case SceneFormat.SceneObject.ObjectMaterialOneofCase.Material:
+                    return ConvertMaterial(parsedObject.Material, id);
                 default:
                     throw new Exception("Unsupported material");
             }
         }
 
-        private Light ConvertToLight(SceneFormat.Light parsedLight)
+        private Light ConvertToLight(SceneFormat.Light parsedLight, int id)
         {
             var tran = ConvertTransform(parsedLight.Transform);
             var color = ConvertColor(parsedLight.Color);
             switch (parsedLight.LightCase)
             {
                 case SceneFormat.Light.LightOneofCase.Point:
-                    return new PointLight(color, 1)
+                    return new PointLight(color, 1, id, SamplerProvider.Sampler)
                     {
                         Id = parsedLight.Id,
                         Transform = tran,
                     };
                 case SceneFormat.Light.LightOneofCase.Directional:
-                    return new DirectionalLight(color, 1)
+                    return new DirectionalLight(color, 1, id, SamplerProvider.Sampler)
                     {
                         Id = parsedLight.Id,
                         Transform = tran,
                     };
                 case SceneFormat.Light.LightOneofCase.Environment:
-                    return new EnvironmentLight(color, 1)
+                    return new EnvironmentLight(color, 1, id, SamplerProvider.Sampler)
                     {
                         Id = parsedLight.Id,
                     };
@@ -166,25 +170,63 @@
             }
         }
 
-        private Dictionary<string, Material> ParseMaterials(RepeatedField<SceneFormat.Material> materials)
+        private Dictionary<string, IMaterial> ParseMaterials(RepeatedField<SceneFormat.Material> materials)
         {
-            var result = new Dictionary<string, Material>();
+            var result = new Dictionary<string, IMaterial>();
             foreach (var material in materials)
             {
-                var mat = ConvertMaterial(material);
+                var mat = ConvertMaterial(material, -1);
                 result[material.Id] = mat;
             }
             return result;
         }
 
-        private Material ConvertMaterial(SceneFormat.Material material)
+        private IMaterial ConvertMaterial(SceneFormat.Material material, int id)
         {
             switch (material.MaterialCase)
             {
                 case SceneFormat.Material.MaterialOneofCase.LambertReflection:
-                    return new DiffuseMaterial(ConvertColor(material.LambertReflection.Color), 1);
+                    return new DiffuseMaterial(ConvertColor(material.LambertReflection.Color),
+                        (float)material.LambertReflection.R, id);
                 case SceneFormat.Material.MaterialOneofCase.SpecularReflection:
-                    return new ReflectionMaterial(1, (float)material.SpecularReflection.Eta);
+                    return new ReflectionMaterial(ConvertColor(material.SpecularReflection.Color),
+                        (float)material.SpecularReflection.R,
+                        (float)material.SpecularReflection.Eta, id);
+                case SceneFormat.Material.MaterialOneofCase.SpecularTransmission:
+                    return new TransmissionMaterial(ConvertColor(material.SpecularTransmission.Color),
+                        (float)material.SpecularTransmission.T,
+                        (float)material.SpecularTransmission.Eta, id);
+                case SceneFormat.Material.MaterialOneofCase.Fresnel:
+                    return new FresnelMaterial(ConvertColor(material.Fresnel.Color),
+                        (float)material.Fresnel.R,
+                        (float)material.Fresnel.T,
+                        (float)material.Fresnel.Eta, id);
+                case SceneFormat.Material.MaterialOneofCase.OrenNayar:
+                    return new OrenNayarMaterial(ConvertColor(material.OrenNayar.Color),
+                        (float)material.OrenNayar.R,
+                        (float)material.OrenNayar.Roughness, id);
+                case SceneFormat.Material.MaterialOneofCase.MicrofacetReflection:
+                    return new MicrofacetReflectionMaterial(ConvertColor(material.MicrofacetReflection.Color),
+                        (float)material.MicrofacetReflection.R,
+                        (float)material.MicrofacetReflection.Eta,
+                        (float)material.MicrofacetReflection.Roughness, id);
+                case SceneFormat.Material.MaterialOneofCase.Metal:
+                    return new MetalMaterial(ConvertColor(material.Metal.Color),
+                        (float)material.Metal.R,
+                        (float)material.Metal.Eta,
+                        (float)material.Metal.K,
+                        (float)material.Metal.Roughness, id);
+                case SceneFormat.Material.MaterialOneofCase.Plastic:
+                    return new PlasticMaterial(ConvertColor(material.Plastic.Color),
+                        (float)material.Plastic.R,
+                        (float)material.Plastic.Roughness, id);
+                case SceneFormat.Material.MaterialOneofCase.Blend:
+                    return new BlendMaterial(ConvertColor(material.Blend.Diffuse),
+                        ConvertColor(material.Blend.Specular),
+                        (float)material.Blend.Roughness, id);
+                case SceneFormat.Material.MaterialOneofCase.Mirror:
+                    return new MirrorMaterial(ConvertColor(material.Mirror.Color),
+                        (float)material.Mirror.R, id);
                 default:
                     throw new Exception("Unsupported material");
             }
